almabase/requirements.txt
appdirs==1.4.0
Django==1.10.5
django-cors-headers==2.0.1
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
mysqlclient==1.3.9
packaging==16.8
PyJWT==1.4.2
pyparsing==2.1.10
six==1.10.0
